feat: released events - requests controller

+++ main-service/src/main/java/ru/practicum/explorewithme/users/dto/ChangeRequestStatusDto.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 @Data
 public class ChangeRequestStatusDto {
-    List<Integer> requestIds;
+    List<Long> requestIds;
     RequestStatus status;
 }
